+++ mobile/src/app/(tabs)/_layout.tsx
