+++ mobile/src/app/(tabs)/two.tsx
@@ -1,7 +1,7 @@
 import { Stack } from 'expo-router';
 import { StyleSheet, View } from 'react-native';
 
-import { ScreenContent } from '~/components/ScreenContent';
+import { ScreenContent } from '~/src/components/ScreenContent';
 
 export default function Home() {
   return (
