+++ mobile/src/components/ScreenContent.tsx
