+++ mobile/tailwind.config.js
@@ -1,6 +1,6 @@
 /** @type {import('tailwindcss').Config} */
 module.exports = {
-  content: ['./index.{js,ts,tsx}', './app/**/*.{js,ts,tsx}', './components/**/*.{js,ts,tsx}'],
+  content: ['./src/index.{js,ts,tsx}', './src/app/**/*.{js,ts,tsx}', './src/components/**/*.{js,ts,tsx}'],
 
   presets: [require('nativewind/preset')],
   theme: {
